getElementBytes: fix the output's format
Setting the width of an std::ostream isn't persistent: it is reset after a
number is put to the stream. As a consequence, it needs to be called before
putting any number in the stream and cannot be factored.

Also, for symmetry with setElementBytes, we now prepend "Ox" to each
hexadecimal value being printed out by getElementBytes, so that when given back
by setElementBytes, it will automatically be parsed as hexadecimal.

Signed-off-by: David Wagner <david.wagner@intel.com>