fix setParameter on array parameters
Since 7ed7046c9e092dbd358b797af678096bc5b692fb, only the first value of an
array could be set by the user and would take the last of the values given by
the user (since each one would overwrite the previous one).

The cause was that in the loop that sets the value, the offset wasn't correctly
incremented.

Signed-off-by: David Wagner <david.wagner@intel.com>