Add a CMake option to not turn warnings into errors
We have some warnings in Release configuration because of:
- unused arguments (some calls to `assert`, that are completly removed in
  release mode, cause some arguments to be unused) in such a pattern:

    void f(bool b)
    {
        assert(b);
    }
- strict aliasing rules violation in floating point tests. Since this is only
  test code, it is not considered critical.

Add an option (on by default) to add the -Werror flag. Have travis travis build
the Parameter Framework in "Release" configuration (it previously built in
"Debug" and "None" configurations) without the -Werror flag. Also, forcefully
remove the "/WX" flag on windows (equivalent to -Werror) because we are not
warning-free yet, even in debug mode.

Signed-off-by: David Wagner <david.wagner@intel.com>