Python bindings: Remove useless nestedworkaround
The nested workaround is used to tell swig to ignore the
inner class definition that would be redundant with the fake outer class.

It would have been useful if ParameterMgrFullConnector.h was included
(as opposed to copying the class definition in this .i).

As their is no conflicting ILogger definition, this workaround is useless.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>