CMake: make the check for XInclude support more robust
We used to use check_library_exists which tries to compile a simple
program involving the required function (xmlXIncludeProcess) and link it
against libxml2. However, this could fail for some other, unrelated reasons.
Even though such reasons could be fatal later, the error message was not clear,
because unrelated to the actual problem.

Instead, we now check that the LIBXML_XINCLUDE_ENABLED macro is defined in
xmlversion.h - which is how libxml2 describes activated features. This does not
involve linking and thus is more robust.

Signed-off-by: David Wagner <david.wagner@intel.com>