functional tests: fix an unused result when trying to display a rich diff
In release builds, with GCC, unused returned values of functions markes as
"warn_unused_results" are enforced. We are calling `system`, that falls in this
category.

We now use the return value and take the opportunity to warn the user that we
didn't manage to use "git diff" to display a rich diff.

Signed-off-by: David Wagner <david.wagner@intel.com>