Allow a client to send "remote" commands programatically
Users can already send commands to a Parameter Framework instance through a
socket (provided that they asked for this interface to be started). Some
features are only available through this interface.

In upcoming patches, we want to use those features and send such commands but
we don't want to use a remote interface, which is sometimes painful. This patch
makes the "remote" commands available through a programmatic API. This API
expects a command name and a string-vector containing the arguments; which
makes it fairly easy to use and to implement a user interface providing the
same commands as they are used to - but without the need of a socket.

The object that provides this API is allocated by the Parameter Framework but
is owned by the client - we want to be C++03-retrocompatible, which forbids us
from returning a unique_ptr, this is why we return a raw pointer.

Signed-off-by: David Wagner <david.wagner@intel.com>
Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>