Don't rely on xsi:noNamespaceSchemaLocation for schema validation
This attribute only makes sense for the XML parser (e.g. libxml2) but since we
don't make libxml2 automatically validate our files (we only do it manually),
we don't need this attribute.

Moreover, this attribute is hard to use:
- when it is set to a relative path, the file containing it cannot be moved;
- when set to an absolute path, it can't be copied between machines without
  modifying that attribute.

We used to set it to a base name ("FooBard.xsd") which was the name of the root
element contatenated with ".xsd" but this is redundant. In tests, it also led
to generating documents with this attribute and a root element that does not
have a corresponding standalone XSD (e.g. ParameterBlock does not have its own
XSD). For instance:

    <ParameterBlock xsi:noNamespaceSchemaLocation="ParameterBlock.xsd">

When trying to validate a file, just rely on the base dir provided by the user
and the root element name in order to find the corresponding XSD file.

Signed-off-by: David Wagner <david.wagner@intel.com>