Remove leading newlines from user-oriented output
They don't make much sense and make it harder to be parsed.

Even though user-oriented output is not meant to be parsed, it is very common
to have user-oriented output be more or less machine-friendly. It actually is
the case of the Parameter Framework's remote commands, except for these leading
newlines.

Signed-off-by: David Wagner <david.wagner@intel.com>