Refactor getListOfElementsWithMapping
The function was implemented using recursive call and method overload.

This was unnecessary complicated and bugged as the subsystem was not
added in the list although it may contain a mapping.

Refactor by using a simple loop and gathering all parents that are
`CConfigurableElement`s.

In order for this method to be reused, change the semantic of the
method to return the node and it's ancestors. It also simplify the
implementation without impact on the rest of the code.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>