Tokenizer: option for not merging delimiters
The tokenizer was ignoring leading and trailing delimiters and also considering
multiple consecutive delimiters as one. This behaviour made it behave similarly
to a shell tokenizer but made it impossible to produce empty tokens.

Another, optional, behaviour is introduced that will produce no token if the
string is empty or N+1 tokens if there are N delimiters.

Signed-off-by: David Wagner <david.wagner@intel.com>