Avoid duplicate or unaligned version information
Currently, we make a commit changing the version number and then we tag it.
This means that for a time, the version number presented by the Parameter
Framework is not aligned with the latest tag.

This commit makes the Parameter Framework get its version information from an
external source - in order of preference:
 - a version forced by the user (for test and release process purposes)
 - information from git git describe)
 - the content of the .version file (which is filed by git archive)
 - or default to v0.0.0-0

Signed-off-by: David Wagner <david.wagner@intel.com>