Switch from coveralls to codecov
Codecov has a nicer UI and more features even though the coverage script is a
bit more painful to use. See https://codecov.io/github/01org/parameter-framework

Among nice new features is the ability to have coverage information displayed
on Pull Requests through a browser extension. On coveralls, it was painful to
find the coverage for a specific revision whereas it is trivial to do on
Codecov.

Signed-off-by: David Wagner <david.wagner@intel.com>