Fix and test the 'showMapping' command
It was segfaulting due to an erroneous static_cast.

Simplify the code related to showMapping by taking into account the fact that
all ConfigurableElements may have a mapping (SystemClass currently doesn't have
any but it can simply return an empty mapping). The leaf element is a special
case because the instantiation mapping may be formatted using "amends" - other
mappings may not.

Signed-off-by: David Wagner <david.wagner@intel.com>