Allow libremote-processor to be compiled without networking support
Introduce an asio stub.

In such a case, libremote-processor becomes useless but still acts as if it was
working correctly. This is an easy way to have a version of libparameter that
does not embed any networking code, which is required in some case for security
compliance reasons.

This can be achieved by turning the NETWORKING option OFF when configuring the
CMake project.

Signed-off-by: David Wagner <david.wagner@intel.com>