Integer template specializations: use builtin types instead of stdint
Template resolution is tied to name mangling. Since we can't know for sure
which builtin type is used for each of int64_t, uint32_t, etc. we can't rely on
these typedefs. Builtin types, however, are reliable.

As an exemple: on some ABIs, size_t can be a long long but uint64_t a mere
long which means that for this particular platform, a call with a size_t
argument will not match a template instantiation with a uint64_t.

Signed-off-by: David Wagner <david.wagner@intel.com>