Use `override` when necessary
Some compilers complain when failing to use the `override` specifier when
overriding a method. clang-tidy can help automatically add it to relevant
methods. It also does a lot of cleanup by removing `virtual` when both
`virtual` and `override` are specified. Same with `override` and `final`: final
methods are implicitly overrides and thus makes the `override` specifier
redundant.

For reference, this was done with clang-tidy-3.8:

    ls parameter/*cpp | xargs -n1 clang-tidy-3.8 \
        -checks=modernize-use-override -fix -p=<path-to-build-dir> \
        -header-filter=parameter

and CMake was ran with the `-DCMAKE_EXPORT_COMPILE_COMMANDS=1` option. Note the
use of xargs: when running clang-tidy on multiple files at once, strange things
happen and the files are left in state that wouldn't compile. clang-format also
needs to be run after clang-tidy.

Signed-off-by: David Wagner <david.wagner@intel.com>