xml-generator: be compatible with both python2.7 and 3.x
The xml-generator scripts were using deprecated Python 2 constructs and were
unable to run with Python 3.x. This is an issue on Windows because CMake seems
to use the highest version available even when there is a version closer to
what was requested. If both 2.7 and 3.x are installed on the system,

    find_package(PythonInterp 2.7 EXACT REQUIRED)

may fail and

    find_package(PythonInterp 2.7 REQUIRED)

may find 3.x.

This is a known issue for some time now:

    https://public.kitware.com/Bug/view.php?id=15661

Signed-off-by: David Wagner <david.wagner@intel.com>