CMake: homogenize how install-time interfaces are propagated
There are two ways of propagating the post-install include requirements: either
by adding an 'INCLUDES DESTINATION' argument to the 'install' directive or by
adding an 'INTERFACE' argument to the 'target_include_directories' and using a
generator expression to make it active at install-time only. This patch prefers
changes for the latter since it is consistent with how build-time requirements
are propagated.

Signed-off-by: David Wagner <david.wagner@intel.com>