IntegerParameterType: Fix Min/Max XML import overflow
In case XML data for Min and Max values are wrongly set (out of range). The
Import function is not checking out of range format. In order to avoid
roll over some additional boundary check need to be done.

For exemple if we have Sign integer on 8 bits and we set Max value to 200
the new code will initialize the Max value to 127.

Signed-off-by: Sebastien Guiriec <sebastien.guiriec@intel.com>